feat(vm): 实现条件跳转指令 JUMP_IF
- 新增 JUMP_IF 指令用于条件跳转
- 定义比较条件常量：CONDITION_EQUAL、CONDITION_NOT_EQUAL 等- 支持多种数据类型的比较（整数、浮点数）
- 实现有符号和无符号数值的比较逻辑- 更新字节码解析器以识别新的指令类型
- 修正浮点数比较中的变量命名错误
- 调整内存页大小类型为 size_t
- 启用程序结束时的返回值输出
- 移除不再使用的 JIT_FOR_RANGE 相关代码

+++ memory.cpp
@@ -38,7 +38,7 @@ namespace lvm
                     {
                         SYSTEM_INFO sysInfo;
                         GetSystemInfo(&sysInfo);
-                        uint64_t pageSize = sysInfo.dwPageSize;
+                        size_t pageSize = sysInfo.dwPageSize;
 
                         auto pageBase = reinterpret_cast<void*>(reinterpret_cast<uint64_t>(faultAddress) & ~(pageSize -
                             1));
@@ -55,7 +55,7 @@ namespace lvm
                     else if (mbi.State == MEM_COMMIT)
                     {
                         // std::cout << "Page is committed but access violation occurred (permissions issue?)" <<
-                            // std::endl;
+                        // std::endl;
                     }
                     else if (mbi.State == MEM_FREE)
                     {
